Implement initial data reader
The decoder.py script reads the raw College Scorecard data. It determines
which information categories have valid data and the data types to use
for each category (ints, floats, or strings). It generates a list
containing the relevant categories, the data type, and the index
within the raw data. A future script will use this information to
build the database.