Improve project file structure
This commit improves the organization of files in the project.
Rather than each of the module being run individually, they are all
imported and ran through the __main__.py module. This will simplify
making changes to the project in the future.

The structure is organized so that the main scripts are in
scorcard-visualizer/ with supplemental files such as the raw data,
temporary files, and the generated database being located in data/.