Create Dbbuilder class
The functions within the dbbuilder module have been rolled into a
Dbbuilder class. This allows the database connection information
to be stored within a Dbbuilder object where it logically belongs.
This makes the module easier to understand and removes the need for
repeatedly calling dbbuilder functions pointing to an outside
sqlite connection/cursor object.

Tests and the main module have been updated to reflect this change.