Add support for multiple y axes on a single plot
When multiple datasets are plotted, the first dataset will use the left
y axis, the second will use the right y axis, and subsequent datasets
will have their y axis added on the right side of the plot. This makes
it easy to plot multiple datasets with different ranges of values to
compare them and study trends.