Update for Jan. 2017 Scorecard raw data
The interface now works with updated College Scorecard data released in January
2017.

The program is now updated for the new default raw data filenames.

Several changes were made to allow for an increased number of data
types. One of the new data types, "ALIAS", could be a string containing
commas, which affected the parsing of the .csv data. This has been dealt
with by replacing the commas in these strings with semicolons before
processing.

Other misc. changes were made to the docstrings, commenting, and general
formatting.